+++ Rubrik/Private/Get-RubrikAPIData.ps1
@@ -1,4 +1,8 @@
-function Get-RubrikAPIData($endpoint) {
+function Get-RubrikAPIData {
+    [cmdletbinding(SupportsShouldProcess)]
+    param(
+        $endpoint
+    )
     <#
         .SYNOPSIS
         Helper function to retrieve API data from Rubrik
@@ -2486,6 +2490,18 @@ function Get-RubrikAPIData($endpoint) {
                 Success     = '202'
             }
         }
+        'Resume-RubrikSLA'                = @{
+            '5.1' = @{
+                Description = 'Resume a new SLA Domain on a Rubrik cluster'
+                URI         = '/api/v2/sla_domain/{id}/pause'
+                Method      = 'Post'
+                Query       = ''
+                Result      = ''
+                Filter      = ''
+                Success     = '200'
+                ObjectTName = 'Rubrik.SLADomain'
+            }
+        }
         'Set-RubrikAvailabilityGroup'           = @{
             '1.0' = @{
                 Description = 'Update a Microsoft SQL availability group.'
@@ -2886,6 +2902,18 @@ function Get-RubrikAPIData($endpoint) {
                 Success     = '200'
             }
         }
+        'Suspend-RubrikSLA'                = @{
+            '5.1' = @{
+                Description = 'Pause a new SLA Domain on a Rubrik cluster'
+                URI         = '/api/v2/sla_domain/{id}/pause'
+                Method      = 'Post'
+                Query       = ''
+                Result      = ''
+                Filter      = ''
+                Success     = '200'
+                ObjectTName = 'Rubrik.SLADomain'
+            }
+        }
         'Set-RubrikVCD'         = @{
             '1.0' = @{
                 Description = 'Updates settings of a vCD connection'
@@ -3259,8 +3287,18 @@ function Get-RubrikAPIData($endpoint) {
         $key = $api.$endpoint.Keys | Sort-Object | Where-Object {[float]$_ -le $ver} | Select-Object -Last 1
     }
 
-    Write-Verbose -Message "Selected $key API Data for $endpoint"
-    # Add the function name to resolve issue #480
-    $api.$endpoint.$key.Add('Function',$endpoint) 
-    return $api.$endpoint.$key
+    if ($null -eq $key) {
+        $ErrorSplat = @{
+            Message = "No matching endpoint found for $EndPoint that corrosponds to the current cluster version."
+            ErrorAction = 'Stop'
+            TargetObject = $api.$endpoint.keys -join ','
+            Category = 'ObjectNotFound'
+        }
+        Write-Error @ErrorSplat
+    } else {
+        Write-Verbose -Message "Selected $key API Data for $endpoint"
+        # Add the function name to resolve issue #480
+        $api.$endpoint.$key.Add('Function',$endpoint) 
+        return $api.$endpoint.$key
+    }
 } # End of function